+++ src/main/java/com/objects/marketbridge/coupon/infra/MemberCouponRepositoryImpl.java
@@ -1,6 +1,7 @@
-package com.objects.marketbridge.product.infra.coupon;
+package com.objects.marketbridge.coupon.infra;
 
 import com.objects.marketbridge.coupon.domain.MemberCoupon;
+import com.objects.marketbridge.coupon.service.port.MemberCouponRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
