+++ src/test/java/com/objects/marketbridge/order/mock/FakeCouponRepository.java
@@ -1,7 +1,7 @@
 package com.objects.marketbridge.order.mock;
 
 import com.objects.marketbridge.coupon.domain.Coupon;
-import com.objects.marketbridge.product.infra.coupon.CouponRepository;
+import com.objects.marketbridge.coupon.service.port.CouponRepository;
 
 import java.util.List;
