+++ src/test/java/com/objects/marketbridge/product/infra/MemberCouponRepositoryImplTest.java
@@ -4,8 +4,8 @@ import com.objects.marketbridge.coupon.domain.Coupon;
 import com.objects.marketbridge.member.domain.Member;
 import com.objects.marketbridge.coupon.domain.MemberCoupon;
 import com.objects.marketbridge.member.service.port.MemberRepository;
-import com.objects.marketbridge.product.infra.coupon.CouponRepository;
-import com.objects.marketbridge.product.infra.coupon.MemberCouponRepository;
+import com.objects.marketbridge.coupon.service.port.CouponRepository;
+import com.objects.marketbridge.coupon.service.port.MemberCouponRepository;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
